+++ frontend/src/common/unixToFormattedTime.ts
@@ -1,3 +1,9 @@
+/**
+ *
+ * @param unixDate epoch/unix date time number
+ * @returns fomatted time in form of HH:MM:SS  (eg, 19:00:00 or 12:10:20)
+ */
+
 export const unixToFormattedTime = (unixDate?: number) => {
   if (!unixDate) {
     return '-'
@@ -7,7 +13,7 @@ export const unixToFormattedTime = (unixDate?: number) => {
   const minutes = '0' + date.getMinutes()
   const seconds = '0' + date.getSeconds()
 
-  const formattedTime = hours + ':' + minutes.substr(-2) + ':' + seconds.substr(-2)
+  const formattedTime = hours + ':' + minutes.substring(-2) + ':' + seconds.substring(-2)
 
   return formattedTime
 }
