+++ frontend/src/components/BookingCard.tsx
@@ -1,11 +1,13 @@
 import React, { useEffect } from 'react'
-import styled from 'styled-components'
+import styled, { css } from 'styled-components'
 import { useParams, useHistory } from 'react-router-dom'
 import { useDispatch, useSelector } from 'react-redux'
 import { RootState } from '../store/types'
 import { PATHS } from '../routes/Routes'
 import { DOMAIN_URL, ENDPOINTS } from '../store/endpoints'
 import { months, days } from '../common/dates'
+import { get24Hourtime } from '../common/get24HourTime'
+import { openUserTelegram } from '../common/telegramMethods'
 
 import messageIcon from '../assets/messageIcon.svg'
 import adminIcon from '../assets/adminIcon.svg'
@@ -25,25 +27,25 @@ const BookingContainer = styled.div`
 const BookingLeftDisplay = styled.div`
   align-self: center;
   width: 80%;
+  font-weight: 600;
 `
 
-const DateLabel = styled.div`
+const DateComponent = styled.div`
   font-style: normal;
-  font-weight: bold;
   font-size: 14px;
   overflow: hidden;
   color: #ff705c;
 `
 
-const TimeLabel = styled.div`
+const TimeComponent = styled.div`
   font-style: normal;
-  font-weight: bold;
   font-size: 14px;
   line-height: 14px;
+  margin: 8px 0px 8px 0px;
 `
+
 const BookingCCAName = styled.div`
   font-style: normal;
-  font-weight: bold;
   font-size: 14px;
   line-height: 14px;
 `
@@ -57,6 +59,7 @@ const BookingRightDisplay = styled.div`
   align-items: center;
   justify-content: center;
 `
+
 const Icon = styled.img`
   padding: 20px;
 `
@@ -75,28 +78,6 @@ export default function BookingCard() {
     selectedFacilityName,
     selectedFacilityId,
   } = useSelector((state: RootState) => state.facilityBooking)
-  const getHumanReadableTime = (eventStartTime: number) => {
-    const date = new Date(eventStartTime * 1000)
-    let hour = date.getHours().toString()
-    if (hour.length == 1) {
-      hour = '0' + hour
-    }
-    let minutes = date.getMinutes().toString()
-    if (minutes.length == 1) {
-      minutes = '0' + minutes
-    }
-
-    return hour + minutes
-  }
-
-  const getHumanReadableDate = (eventTime: number) => {
-    const date = new Date((eventTime + 28800) * 1000)
-    const numday = date.getUTCDate()
-    const monthInt = date.getUTCMonth() + 1
-    const day = date.getUTCDay()
-
-    return numday + '/' + monthInt
-  }
 
   const fetchTelegram = async (booking) => {
     try {
@@ -107,18 +88,14 @@ export default function BookingCard() {
         .then((resp) => resp.json())
         .then((data) => {
           if (data.telegramHandle === '' || data.telegramHandle === undefined) {
-            console.log(data.err)
+            throw data.err
           } else {
-            openTelegram(data.telegramHandle)
+            openUserTelegram(data.telegramHandle)
           }
         })
     } catch (err) {
       console.log(err)
     }
-  }
-  const openTelegram = (userID) => {
-    const site = 'https://telegram.me/' + userID
-    window.open(site)
   }
 
   return (
@@ -126,19 +103,16 @@ export default function BookingCard() {
       {facilityBookings?.map((event) => (
         <BookingContainer
           key={event.bookingID}
-          onClick={() => {
-            history.push(PATHS.VIEW_FACILITY_BOOKING_ID + event.bookingID)
-          }}
+          onClick={() => history.push(PATHS.VIEW_FACILITY_BOOKING_ID + event.bookingID)}
         >
-          {/* <EventAvatar src={dummyAvatar} /> */}
           <BookingLeftDisplay>
-            <DateLabel>
+            <DateComponent>
               {days[ViewStartDate.getDay()]}{' '}
               {ViewStartDate.getDate() + ' ' + months[ViewStartDate.getMonth()] + ' ' + ViewStartDate.getFullYear()}
-            </DateLabel>
-            <TimeLabel>
-              {getHumanReadableTime(event.startTime)} to {getHumanReadableTime(event.endTime)}
-            </TimeLabel>
+            </DateComponent>
+            <TimeComponent>
+              {get24Hourtime(event.startTime)} to {get24Hourtime(event.endTime)}
+            </TimeComponent>
             <BookingCCAName>{event.ccaName ? event.ccaName : 'Personal'}</BookingCCAName>
           </BookingLeftDisplay>
           <BookingRightDisplay>
